FILE-512:move envelope constraints setting to application.conf
